+++ src/App.tsx
@@ -1,25 +1,12 @@
 import React, { ReactElement } from 'react';
-import styled from '@emotion/styled';
-import { useAppDispatch, useUiState, setIsDark } from '@src/store';
+import { Route, Switch } from 'react-router-dom';
 
-const Wrapper = styled.div`
-  ${({ theme }) => theme.flexCol()}
-  background-color: ${({ theme }) => theme.background};
-  .text {
-    color: ${({ theme }) => theme.text};
-  }
-`;
+import Toggle from '@src/components/Toggle';
 
-const App = (): ReactElement => {
-  const dispatch = useAppDispatch();
-  const { isDark } = useUiState();
-
-  return (
-    <Wrapper>
-      <button type="button" onClick={() => dispatch(setIsDark({ isDark: !isDark }))}>{isDark ? 'dark' : 'light'}</button>
-      <div className="text">result</div>
-    </Wrapper>
-  );
-};
+const App = (): ReactElement => (
+  <Switch>
+    <Route exact path="/" component={Toggle} />
+  </Switch>
+);
 
 export default App;
